requirements.txt
aiomysql==0.2.0
alembic==1.11.1
anyio==3.7.0
cfgv==3.3.1
click==8.1.3
distlib==0.3.6
exceptiongroup==1.1.1
fastapi==0.97.0
filelock==3.12.2
greenlet==2.0.2
h11==0.14.0
identify==2.5.24
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
mysqlclient==2.1.1
nodeenv==1.8.0
platformdirs==3.6.0
pre-commit==3.3.3
pydantic==1.10.9
PyMySQL==1.0.3
python-dotenv==1.0.0
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==2.0.16
SQLAlchemy-Utils==0.41.1
starlette==0.27.0
typing_extensions==4.6.3
uvicorn==0.22.0
virtualenv==20.23.1
yarl==1.9.2
